requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cpu
torchvision==0.9.0+cpu
timm==0.5.4
onnxruntime==1.11.0
numpy==1.22.3
Pillow==8.0.1
requests==2.27.1
pyyaml==6.0